requirements.txt
# openai>=1.0.0
openai==0.28
streamlit
pandas
sqlalchemy
python-dotenv